# auth-api/src/main/java/co/com/example/auth/dto/LoginRequest.java
package co.com.example.auth.dto;

import io.swagger.v3.oas.annotations.media.Schema;
import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;

@Schema(name = "LoginRequest")
public record LoginRequest(
        @NotBlank @Email String email,
        @NotBlank String password
) {}